+++ package.json
@@ -8,8 +8,7 @@
     "start:dev": "nodemon",
     "start:prod": "npm run build && node ./dist/main.bundle.cjs",
     "build": "webpack --config ./webpack.config.js --env mode=prod",
-    "dev": "webpack",
-    "fix": "tsc-esm-fix --target='target/es6'",
+    "dev": "webpack --config ./webpack.config.js",
     "test": "jest",
     "test:coverage": "jest --coverage"
   },
@@ -35,7 +34,6 @@
     "clean-webpack-plugin": "^4.0.0",
     "eslint": "^8.31.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
     "jest": "^29.3.1",
     "nodemon": "^2.0.20",
     "ts-jest": "^29.0.5",
@@ -46,6 +44,7 @@
     "webpack-cli": "^5.0.1"
   },
   "dependencies": {
+    "dotenv": "^16.0.3",
     "uuid": "^9.0.0"
   }
 }
