+++ src/constants/PORT.ts
@@ -1,1 +1,5 @@
-export const PORT = process.env.PORT || 4000;
+import { config } from 'dotenv';
+
+config();
+
+export const { PORT } = process.env;
