+++ src/index.ts
@@ -1,7 +1,3 @@
 import { app } from './app/app';
 
 app();
-
-process.on('SIGINT', () => {
-  process.exit();
-});
