+++ src/uses/users.ts
@@ -1,0 +1,3 @@
+import { IUser } from '../types/IUser';
+
+export const users: IUser[] = [];
